api/wwcodesvtools/requirements.txt
Django==3.2.23
django-registration==3.1
django-environ==0.4.5
djangorestframework==3.11.1
pre-commit==2.7.1
djangorestframework-simplejwt==4.6.0
drf-yasg==1.17.1
django-cors-headers==3.5.0
flake8==3.8.4
gunicorn==20.0.4
psycopg2==2.8.5
South==1.0.2
whitenoise==6.3.0